Add docker-registry handling to master
