+++ cluster/juju/layers/kubernetes-common/metadata.yaml
@@ -1,0 +1,3 @@
+requires:
+  docker-registry:
+    interface: docker-registry
