+++ cluster/juju/layers/kubernetes-worker/metadata.yaml
@@ -43,8 +43,6 @@ requires:
     interface: azure-integration
   nfs:
     interface: mount
-  docker-registry:
-    interface: docker-registry
 provides:
   cni:
     interface: kubernetes-cni
